+++ bio_pieces/vcfcat_main.py
@@ -1,6 +1,6 @@
 '''
 Usage:
-    vcfcat filter <FILE1> [ --tag=<TAG> (--ge | --le | --geq | --leq | --eq | --neq) <VALUE> ] [-c]
+    vcfcat filter <FILE1> ( --tag=<TAG> (--ge | --le | --gt | --lt | --eq | --neq) <VALUE> ) [-c]
     vcfcat exists <FILE1> (--tag=<TAG> ) [-c]
     vcfcat ambiguous <FILE1>  [-c]
     vcfcat vcall <FILE1>  [--csv | -c ]
@@ -10,10 +10,28 @@ Usage:
 
 Options:
     -t=<THRESH> --threshold=<THRESH>  Numerical threshold to determine difference [Default: 0]
-    -c, --count  Output the number of records.
-    --no-header  Output without the VCF header info.
-    --csv         Output simplified format
+    -c, --count  Output the number of records instead of the actual file.
+    --csv        Output variant calls in simplified tabular format
+    --gt         Get records Greater Than <VALUE>
+    --ge         Greater than or Equal
+    --lt         Get records Less Thaan <VALUE>
+    --leq        Get records Less than or equal to <VALUE>
+    --eq         Get records Exactly Equal to <VALUE>
+    --neq        Get records Not Equal to <VALUE>
+
+Arguments:
+    filter:  print only vcf records matching the filter as a new vcf file.
+    exists:  only those records where the tag has a value.
+    ambiguous:  only those records where the Called Base is ambiguous.
+    vcall:  those records where the called base and reference base are different.
+    diff:   those records in <FILE1> which were different than the corresponding records (at
+    the same position) in <FILE2> according to some criteria. Providing a --tag argument will show
+    show positions where the files have different entries in that field. The threshold argument (an integer)
+    is used to determine if the values are "different enough", i.e., must be 10 more, or 10 less, etc.
+    stat:   print simple information (like mean, etc.) for all VCF fields in a tabular format.
+    statdiff:   print the (arithmetic) difference between the two file's "stat" output.
 '''
+
 
 from schema import Schema, Use,  Optional
 from docopt import docopt
@@ -21,13 +39,13 @@ import operator
 from bio_pieces import vcfcat
 import sys
 import vcf
-#TODO: figure out diff output
 #TODO: Find a better way to dispatch commandline apps
-ops = ['--ge', '--le', '--geq' , '--leq' , '--eq' , '--neq']
+ops = ['--ge', '--le', '--gt' , '--lt' , '--eq' , '--neq']
 def validate_value(val):
-    if val is None:
+    if val is None or not val.isdigit():
         return val
-    return val if val.isalpha() else int(val)
+    else:
+        return int(val)
 
 def validate_vcf_file(name):
     if name is None:
@@ -38,7 +56,15 @@ def validate_vcf_file(name):
 #               Use(vcf.Reader, error='pyvcf could not read file'),
 #               Use(list))
 
-def run(raw):
+def run(raw_args):
+    '''
+    Validates the arguments by converting the VCF files into lists of VCFRecords, converting
+    threshold to int, and calling validate_value on <VALUE>.
+    returns the command string for dispatching; the new args dictionary after validating, and the first
+    filename so that that file can later be used for a vcf.VCFWriter object.
+    :param dict raw_args: the result of docopt(__doc__)
+    :return str, dict, str: the command string, the new args dict, and the first filename.
+    '''
     commands = ['vcall', 'filter', 'diff', 'stat', 'statdiff', 'exists', 'ambiguous']
     schema_dict=    {
             '<FILE1>' : Use(validate_vcf_file),
@@ -51,12 +77,18 @@ def run(raw):
              #tags.__contains__ #    error='Tag was not valid, should be one of {0}'.format(' '.join(tags)))
         }
     schema_dict.update( dict( (arg, bool) for arg in commands + ops))
-    _args = Schema(schema_dict).validate(raw)
+    _args = Schema(schema_dict).validate(raw_args)
     cmd_str = [k for (k, arg) in _args.items() if k in commands and arg][0]
-    filename = raw['<FILE1>']
+    filename = raw_args['<FILE1>']
     return cmd_str, _args, filename
 
 def dispatch_cmd(cmd_str, args):
+    '''
+    Dispatch the appropriate vcfcat.py function matching the command string (the first commandline argument)
+    :param str cmd_str: the command chosen, aka filter
+    :param dict args: validated and type-converted command-line arguments
+    :return list result: a list of VCFRecords matching the arguments
+    '''
     cmd = getattr(vcfcat, cmd_str)
     if cmd_str == 'filter':
        op = [k for k, arg in args.items() if k in ops and arg ][0]
@@ -82,6 +114,9 @@ def compute():
 
 
 def print_variant_call(rec):
+    '''
+    print vcalls in a simple tabular format
+    '''
     print("{0}\t{1}\t{2}\t{3}".format(
         rec.CHROM, rec.POS, rec.REF, rec.INFO['CB']
     ))